Docs: When to use credentials vs parameters
In our community meeting today, someone asked for an explanation on when
to use credentials vs sensitive parameters. I explained it and then we
all agreed this needed to be a best practice doc.

This explains when to use each, and some finer points on the security
implications of that choice.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>