Backport changes to test bundles
The test bundles (mysql+wordpress) are failing because the helm chart we
were relying on is no longer available. In the v1 branch we updated the
test bundles to not actually install any software, since it made the
tests slow, and we were often runninging into maintenance for them that
wasn't really adding anything to the tests. We just care that the
bundles run and pass along dependency outputs/parameters, etc.

I've updated the test bundles to use the same improvements that we are
using on the release/v1 branch (with tweaks to accommodate v0.38).

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>