Fix race condition in plugin logs routine (#2126)
* Fix race condition in plugin logs routine

When we run a plugin, we start a go routine that captures logs from the
plugin and forwards them to our traceLogger. I've found that there is a
race condition where:

1. We close the plugin
2. The plugin emits a message saying that its closing
3. We cancel the context for the log go routine but ctx.Done doesn't
   immediately trigger on the go routine.
4. Instead the plugin connections's close function returns before the routine
   is stopped.
5. Porter moves on and closes our logger.
6. The log go routine then tries to forward the message to the logger
   which is now closed.

I've added a wait group that ensures that the log go routine finishes
before the plugin connection returns from Close. This ensures that all
resources managed by the plugin connection are cleaned up before Close
returns.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Move when we add to the wait group

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>