Support migration of data from v0 to v1
The porter storage migrate command now supports migrating from v0.38.*
(not previous versions) to v1.0.0-beta.1 and higher.

You must keep your old PORTER_HOME directory around and then create a
new PORTER_HOME that has v1 installed. The storage migrate command reads
the data from the old home, and copies it into the new database for the
v1 version of porter.

Data in the old porter home is not modified at all (and our interface to
interact with it is a readonly subset of the original plugin protocol to
avoid any oopsies).

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>