Support customizing the test porter command
When we run commands in our smoke and integration tests, sometimes it
would help to be able to set an environment variable for just that
command (so the test can still run in parallel), or to change the
directory for a single command.

This exposes the underlying PreparedCommand object that builds the
porter command so that a caller can set extra environment variables,
change the working directory, etc.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>