Update k8s and containerd dependencies
* Update to cnab-go v0.23.4
* Update containerd to v1.6.6
* Updated k8s to v0.24.1.

This does not update docker since buildkit uses a funny unreleased
version of docker. We won't be able to update to a new version of Docker
until there's a release that has the new feature that buildkit relies
upon. See go.mod for a link to the troublesome package in question.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>