Add prow github action
This adds a prow github action that allows specified people (in the
OWNERS file) to comment on a pull request with /lgtm to review the pull
request, or /approve to merge the pull request.

The github action handles executing the commands for you so that you
don't need to have maintainer rights on the repository.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>