Use v1 contributing info on main porter site
Now that all new development should be against the release/v1 branch,
I've updated the contributing tutorial on the main branch to instruct
people to work with release/v1 and use the udpated v1 instructions.

We have been getting a number of confused contributors because they come
to the contributing tutorial from the main site which is for our main
branch and this should help get them the correct info to contribute.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>