Fix timing on logging a plugin connection error
When connecting to a plugin and it fails to connect, log the error
first, and then call close on the connection. This ensures that the
order of the error(s) in our log file/traces match the order the events
happened. Otherwise it's easy to confuse the error from connect with any
additional errors triggered by close, when they are logged out of order.

This fixes the order of the errors being logged which should help with
that.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>