Persist Run.Bundle as string
Since bundle has dollar prefixed fields, we can't call mongo's update on it
and replace the document (which is what I was trying to do during a repeated migration).
I've changed how we store just bundle, so that it's a string field when stored in mongo.
We don't query based on run.Bundle so I don't see this as being a large change. It does change the storage schema version for installations though.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>