Fix yaml tag for dependency version
There was a typo in the yaml tag for a dependency's version in
porter.yaml. It has always been version, and we intended to just move
reference and version underneath a new bundle field. It was accidentally
set to versions, but none of the other code, tests or examples were
updated to that value (yay!)

This reverts the yaml tag back to version (singular). I'm not bumping
the schema version since this actually matches what the schema has
always been, it's just fixing a bug where version completely didn't work
for a bit.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>