Set AllowPrerelease on deps v1 extension
When we are converting a porter.yaml to a bundle.json and populating the
v1 dependencies extension metadata, we accidentally stopped populating
AllowPrerelease (which was in use previously). Before we updated the
dependency schema in preparation for a v2 of the dependency extension
this value was set explicitly in the porter.yaml. When we removed it, we
decided to use the mastermind/semver convention of "if a version
constraint includes a prerelease, then we will match against releases",
i.e. 1.0.0-0 would allow prereleases, but we missed updating the
conversion logic to explicitly set AllowPrerelease in that situation.

I've updated our testdata to work with the new logic and verified that
we are now properly setting AllowPrerelease again.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>