Only load mixin metadata from used mixins (#2214)
* Only load mixin metadata from used mixins

When we build a bundle, we record metadata about each mixin used in the
bundle. Previously we were first retrieving metadata from every single
installed mixin, and then filtering that by the mixins used.

I've changed the build command so that instead we first determine which
mixins to query, and then we only ask those mixins for their metadata.
This avoids problems and unhelpful error messages when installed, but
unused, mixins are queried and return an error. It also just speeds up
build which is nice too.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Review feedback

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>