Fix test flake TestPlugins_CatchStderr
The cause of the inconsistent error message returned when connecting to
a plugin was that there is a bit of a race condition in the
hashicorp/go-plugin framework where messages written to the plugins
stderr may not be reported by the time the plugin framework rejects a
non-conformant message that the plugin wrote to stdout.

The result is that sometimes our code would get a chance to see the
erorr message that the plugin wrote, and sometimes we wouldn't.

I don't want to modify the plugin framework just make one of our tests
more reliable. So instead I have updated our test plugin so that it
waits a bit between writing to stderr and writing to stdout. Now we
reliably get both messages captured by go-plugins.

In addition, I made the error message returned by the plugin connection
consistent regardless of whether or not we were able to capture stderr
from the plugin.

Fixes #2213

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>