Refactor shared bundle action options
Previously we had a shared option struct, sharedOptions and
BundleAction, that were used for a large grouping of commands that work
with bundles, from install to explain.

I've refactored the shared options into more granular groupings so that
later I can define debug mode for _just_ commands that run bundles.
Now there are shared options for commands that execute bundles and commands
that merely reference a bundle (such as explain).

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>