Allow reference as positional arg to explain/inspect
Since the inspect and explain commands operate directly on a bundle, it
makes sense to let the user specify it as a positional argument. I have
updated both commands to accept either the existing --reference flag
(for consistency with other commands and backwards compatibility) and
alow support passing the reference as the first positional argument.

So either of the following now work:

porter explain REFERENCE
porter explain --reference

I've updated the docs to prefer the form without the flag, since it's
shorter.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>