Move move action commands under installations
The bundle action commands (install, invoke, upgrade, uninstall) really
act on installations, not bundles. I am moving those commands under
installations, so that the full command for say "porter install" would
be "porter installation install".

Everyone uses the aliases anyway so I don't expect this to be an
impactful change to the commands people have automated. But where it
does help is better group the commands in the help text. Now you can
find upgrade associated to an installation, which is a better fit.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>