Remove previously generated cli docs before regenerating
When I moved the bundle action commands, this caused some of the
generated cli doc files to no longer exist. But just running mage
docsgen didn't cause those files to be deleted.

I've fixed the DocsGen target so that it first removes the old generated
files, and then regenerates them. That way if a file isn't regenerated,
it is deleted in git.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>