Add --debug flag to bundle actions
The --debug flag is now just for the bundle action commands: install,
upgrade, invoke and uninstall. The flag is used to indicate that the
bundle should run in "debug mode". In debug mode, porter sets the
parameter PORTER_DEBUG to true, and the bundle, porter runtime and
mixins may check this value to have the bundle *behave* differently than
normal. It's not just controlling how verbose the output is.

For example, a mixin may run additional commands to gather debug output,
or what I would like to do with helm3 is have it not use --atomic in
debug mode, so that when a helm release fails, the release is kept
around for troubleshooting.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>