Add context.Context and runtime.RuntimeContext to builder package
Add context plumbing and also the new runtime.RuntimeContext struct
(which has the debug mode flag) to the pkg/exec/builder package which is
what mixins use to implement a lot of their functionality.

This prepares us to get tracing data eventually from mixins when a
bundle runs, and also makes the debug flag available to any mixins who
want to check it.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>