Consolidate --debug, --debug-plugins into --verbosity
The global flags --debug and --debug-plugins have been completely removed
now (this has been an on-going process, this is the last of many PRs
that have incrementally stopped using those flags) and are replaced by a
single global flag: --verbosity. Verbosity controls only the console
output. Log files and traces are controled seperately.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>