Document new location of docker images
A while back we switched to only publishing images to ghcr.io instead of docker.io
due to changes in the Hub service (rate limits, etc). I forgot to update these pages
that document the images when I did that, so this fixes the doc to point to the current
image locations.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>