Fix regression in allow-docker-host-access
I accidentally removed setting allow-docker-host-access from the global
config when the flag was not set in a previous refactoring PR. This
didn't cause the build to fail because there were no regression tests
around that flag.

I've added a regression test for it (hello-test) and at the same time
noticed that the mybuns test bundle had its dependency defined wrong, so
the dependency wasn't running. I fixed that too and updated the mydb
bundle to work when called as a dependency of mybuns.

🥵

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>