Move action commands under installations  (#2255)
* Move move action commands under installations

The bundle action commands (install, invoke, upgrade, uninstall) really
act on installations, not bundles. I am moving those commands under
installations, so that the full command for say "porter install" would
be "porter installation install".

Everyone uses the aliases anyway so I don't expect this to be an
impactful change to the commands people have automated. But where it
does help is better group the commands in the help text. Now you can
find upgrade associated to an installation, which is a better fit.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Remove previously generated cli docs before regenerating

When I moved the bundle action commands, this caused some of the
generated cli doc files to no longer exist. But just running mage
docsgen didn't cause those files to be deleted.

I've fixed the DocsGen target so that it first removes the old generated
files, and then regenerates them. That way if a file isn't regenerated,
it is deleted in git.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fix examples

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fix wonky merge with the --credential-set PR

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>