Remove check for docker socket with --allow-docker-host-access
We were checking for the existence of the docker socket before allowing
the --allow-docker-host-access to be applied. But Docker treats mounting
teh docker socket specially and in some cases, such as on windows or
when connected to a remote docker daemon, it won't be present locally.

I'm removing the check and if there is a problem with granting access to
the docker socket, we'll find out when Docker returns an error when we
start the container.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>