Consolidate --debug, --debug-plugins into --verbosity (#2258)
* Add --debug flag to bundle actions

The --debug flag is now just for the bundle action commands: install,
upgrade, invoke and uninstall. The flag is used to indicate that the
bundle should run in "debug mode". In debug mode, porter sets the
parameter PORTER_DEBUG to true, and the bundle, porter runtime and
mixins may check this value to have the bundle *behave* differently than
normal. It's not just controlling how verbose the output is.

For example, a mixin may run additional commands to gather debug output,
or what I would like to do with helm3 is have it not use --atomic in
debug mode, so that when a helm release fails, the release is kept
around for troubleshooting.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Add context.Context and runtime.RuntimeContext to builder package

Add context plumbing and also the new runtime.RuntimeContext struct
(which has the debug mode flag) to the pkg/exec/builder package which is
what mixins use to implement a lot of their functionality.

This prepares us to get tracing data eventually from mixins when a
bundle runs, and also makes the debug flag available to any mixins who
want to check it.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Update exec mixin to use context and debug mode

This updates the exec mixin to use the context.Context and
runtime.RuntimeContext that was just added to pkg/exec/builder.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Consolidate --debug, --debug-plugins into --verbosity

The global flags --debug and --debug-plugins have been completely removed
now (this has been an on-going process, this is the last of many PRs
that have incrementally stopped using those flags) and are replaced by a
single global flag: --verbosity. Verbosity controls only the console
output. Log files and traces are controled seperately.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Update command documentation

Update the cli documentation to match the updated flags for --debug,
--debug-plugins and --verbosity.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Actually pass DebugMode to porter run

... 🤦‍♀️

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>