Fix support for insecure registries
Two of our dependencies, cnab-to-oci and cnab-go, had implementations of registry communication. cnab-go for working with archives, and cnab-to-oci for pushing bundles to registries and copying bundles. Each had inconsistent or missing logic around when to skip TLS certificate validation or use plain http. So while Porter was passing the --insecure-registry flag to our dependencies, it wasn't having the intended effect.

I have updated both dependencies to support unsecured (plain http) registries on loopback addresses (127.x.x.x) by default, and to skip TLS verification as requested by Porter with the --insecure-registry flag.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>