Use caching in the CI pipeline
Use the Azure DevOps Cache task to cache the go module directory, and the go build directory, so that go build/test run faster. I am using a separate cache key for the native go build (which is mage build) and the cross compiled binaries (which is mage xbuildall), since
they generate different artifacts in the go build directory. Once a cache artifact is persisted, it can't be updated or removed. It is keyed based on our go.sumfile along with the agent OS. When we need to clear the cached artifacts, the leading version prefix should be incremented, which is used just to bust the cache.

This has brought our builds down from ~6m to ~1m per job. The smoke test on Windows is still slow, but I'll address that in another PR (or most likely updating the agent image for Windows).

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>