Add test mixin to test schema validation
I have added a test mixin that returns a static schema that includes a
build time property:

```yaml
mixins:
  - testmixin:
      clientVersion: 1.2.3
```

I've moved the schema test to an integration test, and it is now
validating the standard mybuns test bundle, and also a small invalid
bundle that violates the schema defined by the test mixin.

This checks that when a mixin reports its declaration / configuration
schema, that it is picked up by Porter and can be enforced by the
`porter schema` output used by VS Code.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>