Discard the temporary invocation image tag during publish
When we publish a bundle, Porter creates a temporary tag for the invocation image, pushes it to the tag, and then updates the bundle to reference the invocation image by its digest.

Right now the reference to the image includes both the temporary tag and the digest, e.g.

example.com/mybundle:temporary-tag@sha256:abc123

This creates a long term reference to the temporary tag, which prevents users from pruning it, and the registry's own internal garbage collector from ever allowing it to be removed since the bundle manifest will always reference it.

The tag was intended to be temporary so I have updated publish so that when we determine the repository digest for the invocation image, we rewrite the invocation image reference to use ONLY the digest, and to discard the temporary tag.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>