Make identifiable temporary invocation image tags
When we publish a bundle, we have to push the invocation image temporarily to a tag, so that we can then retrieve its repository digest. Right now the tag is the hash of the bundle reference, which is impossible for someone maintaining a registry to identify and clean up.

This changes the format of the temporary invocation image tag to:

porter-HASH

For example:

The input text for the invocation image hash is the location of the bundle:

ghcr.io/getporter/examples/porter-hello:v0.2.0

which results in a md5 hash of 3cb284ae76addb8d56b52bb7d6838351, so the resulting temporary location where we push the invocation image is:

ghcr.io/getporter/examples/porter-hello:porter-3cb284ae76addb8d56b52bb7d6838351

I have updated all our code to use the same function to generate the temporary tag, even for referenced images which previously used a different method of computing the hash. Referenced images now compute the hash based on the original location of the image.

So if a bundle referenced nginx:v1.0.0 and is used in a bundle located at example.com/mybuns, the temporary location for nginx would be:

examlpe.com/mybuns:porter-55135915b5d3b5cf91b7305ec6a76d6e

where the hash uses the location of the original image (nginx:v1.0.0)

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>