Disable windows smoke tests
Our Windows agent is having awful performance problems because we are
unable to turn off windows defender. This is causing our smoke tests to
take 3-6x longer than it should on Windows, and they are failing with
timeouts in most builds.

Long term we need to build out a new agent that has defender disabled.
Until we have time to do that, I am disabling the windows smoke tests.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>