Clarify in docs that schemaType isn't supported on porter.yaml
I had originally documented that schemaType was supported on porter.yaml
files but then forgot to add it to the manifest data structure (since
it's not actually used, it's just a planned helper field for  VS Code
autocomplete in the future).

When `schemaType: Bundle` is set in porter.yaml, it causes the manifest
to fail validation.

Since this isn't a showstopping problem, more of too eager
documentation, I am updating the documentation for porter.yaml to
clarify that schemaType isn't supported yet.

Post v1, we can fix this bug in a patch.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>