Configure logging when the exec mixin starts
In the main function for exex, I have copied some of the standard main "boilerplate" functionality around initializing a logger, and registering it with the go context. This enables other mixin functions to be able to grab a logger from the context when they call tracing.StartSpan or tracing.LoggerFromContext.

We do not have telemetry properly configured from inside a bundle yet, but this fixes debug logging from runtime libraries called by the mixins. Later on we can complete setting up the mixins so that they can particpate in telemetry/tracing too.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>