Pass porter configuration as env vars to the bundle runtime
When we run a bundle, we need to pass select porter configuration into the bundle using environment variables. This allows us to have the --verbosity flag, or --debug flag, to be used from inside the bundle.

Verbosity affects the log level of what is printed to the console.
DebugMode is used by the mixins to change their behavior when debugging a bundle (such as running slightly different commands).

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>