Replaced tar to improve consistency of archived bundles
The tar implementation used previously created tar file headers which
included OS and user specific values.
Since they couldn't be overwritten the tar file would be very different
when archived on Windows vs Linux - or just be different users in Linux.

This is now replaced by a custom path walk that ensure the tar file
headers are created consistently across platforms and users - which is
possible due to the simple nature of the porter archives.

Consistency is achived by hardcoding the uid/gid/mtime/filemode.

Signed-off-by: Jens Arnfast <jens@arnfast.net>