Pass docker auth when retrieving bundle metadata
Before we publish or copy a bundle, we retrieve the destination bundle's metadata as a way to check if it already exists, and requires --force to overwrite. By default crane passes the current keychain when interacting with registries. Crane uses go-containerregistry under the hood and at the time I didn't realize that it was crane, and not go-containerregistry specifying the authentication. The go-containerregistry library authenticates as anonymous by default. So when a bundle was published to a new location, we would get a 401 auth error because authentication was not supplied, when we expected a 404 to tell that the bundle didn't exist at the destination location.

I have updated the implementation of GetBundleMetadata to use crane for consistency and consolidated the set up of the crane options.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>