Fix macos install check to not use docker
The macos install script check is failing because it calls porter list, which requires docker to be installed on the CI agent. If we had a plugin that didn't require docker configured instead, we could revert this and use that instead. But for now I'm only calling porter version and skipping calling porter list in the macos install check script.

The linux and windows checks will keep running porter list, which vets that our default config works.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>