Use permalink in install scripts for canary
When we upload an installation script to our github releases page for a release, the script is rewritten to include the version of porter to install. When it's a tagged release that works fine, but at the moment it's embedding the version number even when it's not a release, namely the canary build. Canary binaries are not uploaded under v1.0.0-gsomehash, and instead go into a release named "canary".

I've updated the build scripts so that when we upload the fixed install script for a canary release, the permalink "canary" is replaced in the install script instead of the git describe generated version number.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>