Update installation instructions for plugins/mixins
For a while we had to explicitly set the version when installing prerelease mixins/plugins. Now that a bunch of them have 1.0.0 releass, we can remove that since porter will automatically install the most recent stable version.

A couple packages are still using release candidates, so I updated the version number in the install command to the most recent release.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>