Split Installation into a spec and status
I have updated Installation to be composed of
* Instance specific fields that make testing difficult, such as the ID.
* Spec - all fields that a user would set on an installation
* Status - our own internal tracking of the installation status.

This makes it easier for us to test or work with just the values the user specified by having a spec that won't have auto-generated values like IDs, timestamps, etc.
It also enables us to embed an installation spec in another data structure, like a workflow that runs multiple installations, without having unnecessary fields embeded like the status and id. It lets us think about just the desired state of the installation.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>