Organize long smoke tests using subtests
When a smoke test is running it's hard to quickly understand where it failed because they are long and do a lot to avoid costly setup time. I'm thinking that by wrapping logical portions of the larger smoke test into a subtest, that we can get better info out when a test fails on exactly which area failed.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>