Include stack traces when we recover a panic in EndSpan
When we panic inside a span, right now you only end up seeing the panic message, e.g. EOF, but not the associated stack trace.
Because we also recover in main, we get a stack trace on the main function, but it would be much more helpful to get the stack on the originating span.

I've updated our recover in EndSpan so that we record the stack trace when a traced function panics. Eventually we won't need the last resort recover in main, but we aren't quite there yet so the panic is double logged.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>